triton-cv/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
onnx>=1.14.0
nvidia-tensorrt>=8.6.0

